src/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
docker>=7.1.0
cerebras-cloud-sdk>=1.2.0
openai>=1.55.0
pydantic>=2.10.0
python-dotenv>=1.0.1
aiohttp>=3.11.0
rich>=13.9.0
tenacity>=9.0.0